test(suite-desktop): split connect-ws.test file per connect method
